requirements.txt
django==4.2.4
djangorestframework==3.14.0
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.23.5
keras==2.13.1
tensorflow==2.13
Pillow==7.2.0